refactor rearrange functions in files
